Refactor: move common functions out of csr controller, into `constants` or `helpers`.
